Backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pymongo==4.8.0
pydantic==2.9.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
httpx==0.27.0
slowapi==0.1.9
google-ai-generativelanguage==0.6.15
google-auth==2.38.0
google-api-python-client==2.166.0
fastapi-utils==0.2.1